Format Nullable!string manually
This is a workaround for a phobos bug where putting Nullable!string in
an array and converting the wrapping array to a string with format() or
to!string() causes an assertion error if Nullable!string is null. We
print Nullables in error messages, so mocked functions that have
Nullable!string arguments fail with an assertion instead of the
approprate exception.

We use format!"%(%s %)"([arg]) to print the argument in quotes if arg is
a string. This commit handles strings separately, so there is no need to
put Nullable!string argument into an array. There is still 1 place
causing a deprecation warning, but it is internal to std and doesn't
seem to break anything so far: Nullable!(Tuple!(Nullable!string)) in
meta.d.

Fixes #19.